asd
asd

+++ KatamariDamacy/Client/Codes/TimeUI.cpp
@@ -43,7 +43,7 @@ HRESULT CTimeUI::Initialize_Clone(void * pArg)
 	m_TransformMatrix._41 = 550.f;
 	m_TransformMatrix._42 = 270.f;
 
-	XMStoreFloat4x4(&m_OrthMatrix, XMMatrixOrthographicLH(g_iWinCX, g_iWinCY, 0.0f, 1.f));
+	XMStoreFloat4x4(&m_OrthMatrix, XMMatrixOrthographicLH((float)g_iWinCX, (float)g_iWinCY, 0.0f, 1.f));
 
 	return S_OK;
 }
